requirements.txt
boto3==1.34.0
botocore==1.34.0
click==8.1.7
pyyaml==6.0.1
jinja2==3.1.2
tabulate==0.9.0
colorama==0.4.6
python-dotenv==1.0.0